==> 5_DT_Explain.ipynb <==
# XAI Decision Tree

# Caricamento dei dati di addestramento e salvataggio delle classi e delle feature in diverse variabili
train = pd.read_csv(f"XAI_Datasets/XAI_DoubleFeature/Train_XAI_DoubleFeature.tsv", sep='\t') 
y_train = np.array(train['Order_Label'])
X_train = train.drop('Order_Label', axis=1)

# Caricamento dei dati di test e salvataggio delle classi e delle feature in diverse variabili
test= pd.read_csv(f"XAI_Datasets/XAI_DoubleFeature/Test_XAI_DoubleFeature.tsv", sep='\t')  
y_test = np.array(test['Order_Label'])
X_test = test.drop('Order_Label', axis=1)

# Impostazione del nome della classe e dei nomi delle feature
class_name = 'Order_Label'
feature_names = X_train.columns.tolist()

# Definizione ed addestramento del modello Random Forest
bb = RandomForestClassifier()
bb.fit(X_train, y_train)

# Funzione per effettuare predizioni delle classi 
def bb_predict(X):
    return bb.predict(X)
# Funzione per estrarre le probabilità predette associate alle classi di output
def bb_predict_proba(X):
    return bb.predict_proba(X)

# Valutazione dell'accuratezza 
y_pred = bb_predict(X_test)
print('Accuracy %.3f' % accuracy_score(y_test, y_pred))
# Stampa del classification report del Random Forest
print(classification_report(y_test, y_pred))

# Realizzazione predizioni sul set di addestramento
y_pred_train = bb_predict(X_train)

# Definizione del DecisionTreeClassifier come algorimto di spiegabilità
explainer = DecisionTreeClassifier(max_depth=4)
explainer.fit(X_train, y_pred_train)

# Visualizzazione di come un decision tree approssima il comportamento del modello Random Forest
dot_data = tree.export_graphviz(explainer, out_file=None,  
                                feature_names=feature_names, 
                                class_names=['Inverse_Order' if x == 0 else 'Correct_order' for x in explainer.classes_],
                                filled=True, rounded=True,  
                                special_characters=True)  

graph = pydotplus.graph_from_dot_data(dot_data)  
# Salvataggio e display della rappresentazione 
graph.write_png("DECISION TREE_DoubleFeature_4.png")
Image(graph.create_png())
